plugin: extract PackAccessor into a shared header
The eval store needs the same accessor.
